Prefixed all BIRT POJO data source related classes with "XXX-"
This was done so that classes with the original names would be loaded
from birt-report-tests-with-dependencies.jar, which was uploaded as a
JAR asset; otherwise, the classes from this "report-server" project were
loaded instead.

I also commented out the setting "<attachClasses>true</attachClasses>"
in pom.xml, since it is no longer needed.